+++ extension-style/src/main/java/com/mapbox/maps/extension/style/layers/CustomLayer.kt
@@ -1,0 +1,345 @@
+package com.mapbox.maps.extension.style.layers
+
+import androidx.annotation.UiThread
+import com.mapbox.bindgen.Expected
+import com.mapbox.bindgen.None
+import com.mapbox.bindgen.Value
+import com.mapbox.maps.CustomLayerHost
+import com.mapbox.maps.LayerPosition
+import com.mapbox.maps.MapboxExperimental
+import com.mapbox.maps.Style
+import com.mapbox.maps.StyleManager
+import com.mapbox.maps.extension.style.expressions.generated.Expression
+import com.mapbox.maps.extension.style.layers.properties.PropertyValue
+import com.mapbox.maps.extension.style.layers.properties.generated.Visibility
+import com.mapbox.maps.extension.style.types.LayersDsl
+import com.mapbox.maps.extension.style.utils.silentUnwrap
+import java.util.Locale
+
+/**
+ * Layer with custom rendering implementation.
+ *
+ * With the power of [CustomLayerHost] you can implement your own Open GL ES rendering behaviour
+ * and manipulate layer as a usual one.
+ *
+ * @param layerId mandatory custom layer id.
+ * @param host an instance of user-defined OpenGL ES renderer.
+ *  Note: there is no API to get instance of [CustomLayerHost] from [CustomLayer] after setting it here so
+ *  this instance should be cached on user's end if needed.
+ */
+@UiThread
+class CustomLayer constructor(
+  override val layerId: String,
+  private val host: CustomLayerHost,
+) : Layer(), CustomLayerDsl {
+
+  /**
+   * Get the type of this layer
+   *
+   * @return Type of the layer as [String]
+   */
+  override fun getType(): String {
+    return "custom"
+  }
+
+  /**
+   * Whether this layer is displayed.
+   */
+  override val visibility: Visibility?
+    /**
+     * Whether this layer is displayed.
+     *
+     * Use static method [CustomLayer.defaultVisibility] to get the default property value.
+     *
+     * @return VISIBILITY
+     */
+    get() {
+      val property: String? = getPropertyValue("visibility")
+      property?.let {
+        return Visibility.valueOf(it.uppercase(Locale.US).replace('-', '_'))
+      }
+      return null
+    }
+
+  /**
+   * Whether this layer is displayed.
+   */
+  override val visibilityAsExpression: Expression?
+    /**
+     * Whether this layer is displayed.
+     *
+     * Use static method [CustomLayer.defaultVisibility] to get the default property value.
+     *
+     * @return VISIBILITY as expression
+     */
+    get() {
+      getPropertyValue<Expression>("visibility")?.let {
+        return it
+      }
+      return null
+    }
+
+  /**
+   * Whether this layer is displayed.
+   *
+   * Use static method [CustomLayer.defaultVisibility] to get the default property value.
+   *
+   * @param visibility value of Visibility
+   */
+  override fun visibility(visibility: Visibility): CustomLayer = apply {
+    val propertyValue = PropertyValue("visibility", visibility)
+    setProperty(propertyValue)
+  }
+
+  /**
+   * Whether this layer is displayed.
+   *
+   * Use static method [[CustomLayer.defaultVisibility] to get the default property value.
+   *
+   * @param visibility value of Visibility
+   */
+  override fun visibility(visibility: Expression): CustomLayer = apply {
+    val propertyValue = PropertyValue("visibility", visibility)
+    setProperty(propertyValue)
+  }
+
+  /**
+   * The minimum zoom level for the layer. At zoom levels less than the minzoom, the layer will be hidden.
+   *
+   * Range:
+   *       minimum: 0
+   *       maximum: 24
+   */
+  override val minZoom: Double?
+    /**
+     * Get the minzoom property
+     *
+     * Use static method [CustomLayer.defaultMinZoom] to get the default property value.
+     *
+     * @return minzoom
+     */
+    get() {
+      return getPropertyValue("minzoom")
+    }
+
+  /**
+   * The minimum zoom level for the layer. At zoom levels less than the minzoom, the layer will be hidden.
+   *
+   * Range:
+   *       minimum: 0
+   *       maximum: 24
+   *
+   * Use static method [CustomLayer.defaultMinZoom] to get the default property value.
+   *
+   * @param minZoom value of minzoom
+   */
+  override fun minZoom(minZoom: Double): CustomLayer = apply {
+    val param = PropertyValue("minzoom", minZoom)
+    setProperty(param)
+  }
+
+  /**
+   * The maximum zoom level for the layer. At zoom levels equal to or greater than the maxzoom, the layer will be hidden.
+   *
+   * Range:
+   *       minimum: 0
+   *       maximum: 24
+   */
+  override val maxZoom: Double?
+    /**
+     * Get the maxzoom property
+     *
+     * Use static method [CustomLayer.defaultMaxZoom] to get the default property value.
+     *
+     * @return maxzoom
+     */
+    get() {
+      return getPropertyValue("maxzoom")
+    }
+
+  /**
+   * The maximum zoom level for the layer. At zoom levels equal to or greater than the maxzoom, the layer will be hidden.
+   *
+   * Range:
+   *       minimum: 0
+   *       maximum: 24
+   *
+   * Use static method [CustomLayer.defaultMaxZoom] to get the default property value.
+   *
+   * @param maxZoom value of maxzoom
+   */
+  override fun maxZoom(maxZoom: Double): CustomLayer = apply {
+    val param = PropertyValue("maxzoom", maxZoom)
+    setProperty(param)
+  }
+
+  /**
+   * The slot this layer is assigned to. If specified, and a slot with that name exists,
+   * it will be placed at that position in the layer order.
+   *
+   * @param slot value of slot
+   */
+  @MapboxExperimental
+  override fun slot(slot: String): CustomLayer = apply {
+    val param = PropertyValue("slot", slot)
+    setProperty(param)
+  }
+
+  /**
+   * The slot this layer is assigned to. If specified, and a slot with that name exists,
+   * it will be placed at that position in the layer order.
+   */
+  @MapboxExperimental
+  override val slot: String?
+    /**
+     * Get the slot property
+     *
+     * @return slot
+     */
+    get() {
+      return getPropertyValue("slot")
+    }
+
+  override fun addPersistentLayer(
+    delegate: Style,
+    position: LayerPosition?
+  ): Expected<String, None> {
+    val customLayer = delegate.addPersistentStyleCustomLayer(layerId, host, position)
+    delegate.setStyleLayerProperties(layerId, getCachedLayerProperties())
+    return customLayer
+  }
+
+  override fun addLayer(
+    delegate: Style,
+    propertiesValue: Value,
+    position: LayerPosition?
+  ): Expected<String, None> {
+    val customLayer = delegate.addStyleCustomLayer(layerId, host, position)
+    delegate.setStyleLayerProperties(layerId, propertiesValue)
+    return customLayer
+  }
+
+  /**
+   * Static methods and variables.
+   */
+  companion object {
+    // Default values for layer properties
+    /**
+     * Visibility of the layer.
+     */
+    val defaultVisibility: Visibility?
+      /**
+       * Get the default Visibility property
+       *
+       * @return VISIBILITY
+       */
+      get() {
+        StyleManager.getStyleLayerPropertyDefaultValue("custom", "visibility")
+          .silentUnwrap<String>()?.let {
+            return Visibility.valueOf(it.uppercase(Locale.US).replace('-', '_'))
+          }
+        return null
+      }
+
+    /**
+     * The minimum zoom level for the layer. At zoom levels less than the minzoom, the layer will be hidden.
+     *
+     * Range:
+     *       minimum: 0
+     *       maximum: 24
+     */
+    val defaultMinZoom: Double?
+      /**
+       * Get the minzoom property
+       *
+       * @return minzoom
+       */
+      get() = StyleManager.getStyleLayerPropertyDefaultValue("custom", "minzoom").silentUnwrap()
+
+    /**
+     * The maximum zoom level for the layer. At zoom levels equal to or greater than the maxzoom, the layer will be hidden.
+     *
+     * Range:
+     *       minimum: 0
+     *       maximum: 24
+     */
+    val defaultMaxZoom: Double?
+      /**
+       * Get the maxzoom property
+       *
+       * @return maxzoom
+       */
+      get() = StyleManager.getStyleLayerPropertyDefaultValue("custom", "maxzoom").silentUnwrap()
+  }
+}
+
+/**
+ * This Interface contains all the functions that will be exposed to Kotlin DSL.
+ *
+ * Separated the DSL receiver class to this interface to avoid IDE code suggestion for
+ * property getters.
+ */
+@LayersDsl
+interface CustomLayerDsl {
+
+  /**
+   * The slot this layer is assigned to. If specified, and a slot with that name exists,
+   * it will be placed at that position in the layer order.
+   *
+   * @param slot value of slot
+   */
+  @MapboxExperimental
+  fun slot(slot: String): CustomLayer
+
+  /**
+   * Whether this layer is displayed.
+   *
+   * @param visibility value of Visibility
+   */
+  fun visibility(visibility: Visibility): CustomLayer
+
+  /**
+   * Whether this layer is displayed.
+   *
+   * @param visibility value of Visibility as Expression
+   */
+  fun visibility(visibility: Expression): CustomLayer
+
+  /**
+   * The minimum zoom level for the layer. At zoom levels less than the minzoom, the layer will be hidden.
+   *
+   * Range:
+   *       minimum: 0
+   *       maximum: 24
+   *
+   * @param minZoom value of minzoom
+   */
+  fun minZoom(minZoom: Double): CustomLayer
+
+  /**
+   * The maximum zoom level for the layer. At zoom levels equal to or greater than the maxzoom, the layer will be hidden.
+   *
+   * Range:
+   *       minimum: 0
+   *       maximum: 24
+   *
+   * @param maxZoom value of maxzoom
+   */
+  fun maxZoom(maxZoom: Double): CustomLayer
+}
+
+/**
+ * DSL function for creating a [CustomLayer].
+ */
+@JvmOverloads
+fun customLayer(
+  layerId: String,
+  host: CustomLayerHost,
+  block: (CustomLayerDsl.() -> Unit)? = null
+): CustomLayer {
+  return if (block != null) {
+    CustomLayer(layerId, host).apply(block)
+  } else {
+    CustomLayer(layerId, host)
+  }
+}